Added button schematic
The button will be on a separate PCB so that it can be placed at the
other side of the cube.

The button includes a capacitor for debounce and an extra resistor as a
precaution, incase the button is ever connected to an output*.

*Should be safe for up to 5v, passing 20ma which is pretty common for
MCU I/O. Requires atleast a 1/10W dissipation.